+++ R/RcppExports.R
@@ -25,10 +25,6 @@ LogNorm <- function(data, scale_factor, display_progress = TRUE) {
     .Call('_Seurat_LogNorm', PACKAGE = 'Seurat', data, scale_factor, display_progress)
 }
 
-FastRowScale <- function(mat, scale = TRUE, center = TRUE, scale_max = 10, display_progress = TRUE) {
-    .Call('_Seurat_FastRowScale', PACKAGE = 'Seurat', mat, scale, center, scale_max, display_progress)
-}
-
 Standardize <- function(mat, display_progress = TRUE) {
     .Call('_Seurat_Standardize', PACKAGE = 'Seurat', mat, display_progress)
 }
@@ -79,6 +75,10 @@ SparseRowVar <- function(mat, display_progress) {
 
 ReplaceColsC <- function(mat, col_idx, replacement) {
     .Call('_Seurat_ReplaceColsC', PACKAGE = 'Seurat', mat, col_idx, replacement)
+}
+
+GraphToNeighborHelper <- function(mat) {
+    .Call('_Seurat_GraphToNeighborHelper', PACKAGE = 'Seurat', mat)
 }
 
 fast_dist <- function(x, y, n) {