+++ src/data_manipulation.cpp
@@ -4,6 +4,7 @@
 #include <unordered_map>
 #include <fstream>
 #include <string>
+#include <Rinternals.h>
 
 using namespace Rcpp;
 // [[Rcpp::depends(RcppEigen)]]
@@ -172,40 +173,6 @@ Eigen::SparseMatrix<double> LogNorm(Eigen::SparseMatrix<double> data, int scale_
     }
   }
   return data;
-}
-
-/* Performs row scaling and/or centering. Equivalent to using t(scale(t(mat))) in R.
-   Note: Doesn't handle NA/NaNs in the same way the R implementation does, */
-
-// [[Rcpp::export]]
-Eigen::MatrixXd FastRowScale(Eigen::MatrixXd mat, bool scale = true, bool center = true,
-                             double scale_max = 10, bool display_progress = true){
-  Progress p(mat.rows(), display_progress);
-  Eigen::MatrixXd scaled_mat(mat.rows(), mat.cols());
-  for(int i=0; i < mat.rows(); ++i){
-    p.increment();
-    Eigen::ArrayXd r = mat.row(i).array();
-    double rowMean = r.mean();
-    double rowSdev = 1;
-    if(scale == true){
-      if(center == true){
-        rowSdev = sqrt((r - rowMean).square().sum() / (mat.cols() - 1));
-      }
-      else{
-        rowSdev = sqrt(r.square().sum() / (mat.cols() - 1));
-      }
-    }
-    if(center == false){
-      rowMean = 0;
-    }
-    scaled_mat.row(i) = (r - rowMean) / rowSdev;
-    for(int s=0; s<scaled_mat.row(i).size(); ++s){
-      if(scaled_mat(i, s) > scale_max){
-        scaled_mat(i, s) = scale_max;
-      }
-    }
-  }
-  return scaled_mat;
 }
 
 /* Performs column scaling and/or centering. Equivalent to using scale(mat, TRUE, apply(x,2,sd)) in R.
@@ -496,4 +463,53 @@ Eigen::SparseMatrix<double> ReplaceColsC(Eigen::SparseMatrix<double> mat, Numeri
   return(mat);
 }
 
+template <typename S>
+std::vector<size_t> sort_indexes(const std::vector<S> &v) {
+  // initialize original index locations
+  std::vector<size_t> idx(v.size());
+  std::iota(idx.begin(), idx.end(), 0);
+  std::stable_sort(idx.begin(), idx.end(),
+                   [&v](size_t i1, size_t i2) {return v[i1] < v[i2];});
+  return idx;
+}
+
+//[[Rcpp::export]]
+List GraphToNeighborHelper(Eigen::SparseMatrix<double> mat) {
+  mat = mat.transpose();
+  //determine the number of neighbors
+  int n = 0;
+  for(Eigen::SparseMatrix<double>::InnerIterator it(mat, 0); it; ++it) {
+    n += 1;
+  }
+  Eigen::MatrixXd nn_idx(mat.rows(), n);
+  Eigen::MatrixXd nn_dist(mat.rows(), n);
+  
+  for (int k=0; k<mat.outerSize(); ++k){
+    int n_k = 0;
+    std::vector<double> row_idx;
+    std::vector<double> row_dist;
+    row_idx.reserve(n);
+    row_dist.reserve(n);
+    for (Eigen::SparseMatrix<double>::InnerIterator it(mat,k); it; ++it) {
+      if (n_k > (n-1)) {
+        Rcpp::stop("Not all cells have an equal number of neighbors.");
+      }
+      row_idx.push_back(it.row() + 1);
+      row_dist.push_back(it.value());
+      n_k += 1;
+    }
+    if (n_k != n) {
+      Rcpp::Rcout << n << ":::" << n_k << std::endl;
+      Rcpp::stop("Not all cells have an equal number of neighbors.");
+    }
+    //order the idx based on dist
+    std::vector<size_t> idx_order = sort_indexes(row_dist);
+    for(int i = 0; i < n; ++i) {
+      nn_idx(k, i) = row_idx[idx_order[i]];
+      nn_dist(k, i) = row_dist[idx_order[i]];
+    }
+  }
+  List neighbors = List::create(nn_idx, nn_dist);
+  return(neighbors);
+}